Caching the Inverse of a Matrix
It has two functions, one to create the list + functions
to get and set the cached matrix inverse. The other to
access call those functions.

The two functions then operate together. The first
to create the list + function structure, the other to use it
and compute/return the cached inverse matrix result.